refactor: 패키지 구조 수정, 500에러 알림 aop로 분리&dev알림클래스 삭제
