+++ .gitignore
@@ -37,4 +37,6 @@ out/
 .vscode/
 
 application.yml
+application-dev.yml
+application-prod.yml
 \n# Ignore environment-specific yml files\napplication-*.yml\n!application.yml
