+++ src/main/java/com/example/umc9th/global/notification/NotificationType.java
@@ -1,0 +1,5 @@
+package com.example.umc9th.global.notification;
+
+public enum NotificationType {
+    DISCORD, SLACK
+}
