+++ src/main/java/com/example/umc9th/global/notification/Notifier.java
@@ -1,5 +1,6 @@
-package com.example.umc9th.config.discord;
+package com.example.umc9th.global.notification;
 
 public interface Notifier {
+    NotificationType getType();
     void sendNotification(Exception e, String requestUri);
 }
